refactor: Main, UserGenerator 패키지 이동
