+++ src/main/java/ladder/util/UserGenerator.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.util;
 
 import ladder.domain.User;
