+++ src/test/java/ladder/UserGeneratorTest.java
@@ -1,6 +1,7 @@
 package ladder;
 
 import ladder.domain.User;
+import ladder.util.UserGenerator;
 import org.junit.jupiter.api.Test;
 import java.util.List;
 import static org.assertj.core.api.Assertions.*;
